+++ src/com/nixinova/world/Block.java
@@ -2,6 +2,8 @@ package com.nixinova.world;
 
 import com.nixinova.graphics.Render;
 import com.nixinova.graphics.Textures;
+import com.nixinova.readwrite.Options;
+import com.nixinova.types.BlockCoord;
 
 public class Block {
 	public static Block SKY = new Block("env/sky");
@@ -20,6 +22,13 @@ public class Block {
 	
 	public Render getTexture() {
 		return this.texture;
+	}
+	
+	public static boolean isInsideWorld(BlockCoord coords) {
+		int x = coords.x;
+		int z = coords.z;
+		int size = Options.worldSize;
+		return x >= -size && x <= size && z >= -size && z <= size;
 	}
 
 }